requirements.txt
annotated-types==0.7.0
anyio==4.12.1
babel==2.18.0
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.4
courlan==1.3.2
dateparser==1.3.0
feedparser==6.0.12
h11==0.16.0
htmldate==1.9.4
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jusText==3.0.2
lxml==6.0.2
lxml_html_clean==0.4.4
pdfminer.six==20260107
striprtf==0.0.29
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2025.2
regex==2026.2.28
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.8.3
SQLAlchemy==2.0.48
tld==0.13.1
trafilatura==2.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
